+++ AcountService/entity/Discount.cs
@@ -5,7 +5,7 @@
         public int DiscountId { get; set; }
         public string code { get; set; }
 
-        public double Percent { get; set; }
+        public decimal Percent { get; set; }
